fix contentful routing
